fix: root layout에 use client 제거
